PBL/AIML/Code/requirements.txt
streamlit
requests
numpy
scikit-learn
python-dotenv